// SDKs/twilio.com/twilio_numbers_v2/1.37.3/typescript/src/sdk/models/shared/numbersv2regulatorycompliancebundleitemassignment.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";



export class NumbersV2RegulatoryComplianceBundleItemAssignment extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @SpeakeasyMetadata({ data: "json, name=bundle_sid" })
  bundleSid?: string;

  @SpeakeasyMetadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @SpeakeasyMetadata({ data: "json, name=object_sid" })
  objectSid?: string;

  @SpeakeasyMetadata({ data: "json, name=sid" })
  sid?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}